refactor: rename file with filterUndefinedProps methods
